CXStruct.go: Collect the definition of CXStruct and all methods into one file.
